umkm fitur di admin

+++ database/migrations/2025_07_01_074153_add_opening_hours_to_umkms_table.php
@@ -6,24 +6,22 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('umkms', function (Blueprint $table) {
-            // Tambah kolom jam buka
-            $table->json('opening_hours')->nullable()->after('facebook');
-        });
+        // Cek dulu apakah kolom sudah ada
+        if (!\Schema::hasColumn('umkms', 'opening_hours')) {
+            Schema::table('umkms', function (Blueprint $table) {
+                $table->json('opening_hours')->nullable()->after('facebook');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('umkms', function (Blueprint $table) {
-            $table->dropColumn('opening_hours');
+            if (\Schema::hasColumn('umkms', 'opening_hours')) {
+                $table->dropColumn('opening_hours');
+            }
         });
     }
 };
